Remove SQLAlchemy, add metrics and fix authentication system.
Removed SQLAlchemy and replaced it with asyncpg. Added metrics with
Prometheus and Grafana. Fixed the authentication system and added a new
endpoint to receive refrsh token. API has now both refresh and access
tokens. Backend is currently relying on raw SQL rather than an ORM.

New endpoint /login to receive refresh token and store it in
cookies in /showtoken. Prometheus and grafana are handling the metrics.
Update docker compose and dockerfile to include the new services.
Prometheus uses possibly slowing down code in Middlweware, requires
testing to confirm.

Includes general fixes and improvements, this is not ready to be pushed
to main branch.